feat(ttj): strange tsc

+++ packages/ts-types-json/build.sh
@@ -2,7 +2,10 @@ rm -rf dist/
 
 tsc --declaration
 
-mv dist/src/* dist/
+if [ -e dist/src ] 
+then
+  mv dist/src/* dist/
 
-rm -r dist/src
-rm dist/tsconfig.json
+  rm -r dist/src
+  rm dist/tsconfig.json
+fi
